Correction of explained variance calculation
As proposed by Reinoud Maex in his pull request, the computation of explained variance was incorrect in KeyPy. If we look back to the source of this equation, Pascual-Marqui 1995, reducing equations 10, 11, and 12, leaves a square of the product of the data and model. KeyPy originally computed this product, but did not square it. This fixes that issue.